+++ R/bayes_R2.R
@@ -19,7 +19,7 @@
 #' @references Andrew Gelman, Ben Goodrich, Jonah Gabry & Aki Vehtari. (2019).
 #'   R-squared for Bayesian regression models, \emph{The American Statistician},
 #'   73(3):307-309. \code{10.1080/00031305.2018.1549100} (Preprint available at
-#'   \url{https://stat.columbia.edu/~gelman/research/published/bayes_R2_v3.pdf})
+#'   \url{https://acris.aalto.fi/ws/portalfiles/portal/34206843/bayes_R2_v3.pdf})
 #'
 #' @examples
 #' \dontrun{