+++ R/family-lists.R
@@ -223,7 +223,6 @@
     ybounds = c(0, Inf), closed = c(TRUE, NA),
     ad = c("weights", "subset", "cens", "trunc", "rate", "index"),
     specials = "sbi_log",
-    # experimental use of default priors stored in families #1614
     prior = function(dpar, link = "identity", ...) {
       if (dpar == "shape" && link == "identity") {
         return("inv_gamma(0.4, 0.3)")
@@ -391,7 +390,6 @@
     ad = c("weights", "subset", "cens", "trunc", "mi", "index"),
     include = c("fun_tan_half.stan"),
     normalized = "",
-    # experimental use of default priors stored in families #1614
     prior = function(dpar, link = "identity", ...) {
       if (dpar == "mu" && link == "tan_half") {
         return("student_t(1, 0, 1)")
@@ -513,7 +511,13 @@
     ad = c("weights", "subset", "cens", "trunc", "index"),
     include = "fun_hurdle_negbinomial.stan",
     specials = c("sbi_log", "sbi_hu_logit"),
-    normalized = ""
+    normalized = "",
+    prior = function(dpar, link = "identity", ...) {
+      if (dpar == "shape" && link == "identity") {
+        return("inv_gamma(0.4, 0.3)")
+      }
+      NULL
+    }
   )
 }
 
@@ -578,7 +582,13 @@
     ad = c("weights", "subset", "cens", "trunc", "index"),
     include = "fun_zero_inflated_negbinomial.stan",
     specials = c("sbi_log", "sbi_zi_logit"),
-    normalized = ""
+    normalized = "",
+    prior = function(dpar, link = "identity", ...) {
+      if (dpar == "shape" && link == "identity") {
+        return("inv_gamma(0.4, 0.3)")
+      }
+      NULL
+    }
   )
 }