+++ R/priors.R
@@ -553,6 +553,14 @@ default_prior.default <- function(object, data, family = gaussian(), autocor = N
   .default_prior(bframe, ...)
 }
 
+#' @export
+default_prior.brmsfit <- function(object, ...) {
+  object <- restructure(object)
+  bterms <- brmsterms(object$formula)
+  bframe <- brmsframe(bterms, data = object$data)
+  .default_prior(bframe, ...)
+}
+
 # internal work function of 'default_prior'
 # @param internal return priors for internal use?
 # @return a brmsprior object
@@ -1091,7 +1099,7 @@ def_dpar_prior <- function(x, dpar) {
   dpar_class <- dpar_class(dpar, family = x)
   link <- x$dpars[[dpar]]$family$link %||% "identity"
   if (is.function(x$family$prior)) {
-    # experimental use of default priors stored in families #1614
+    # use default priors stored in families #1614
     # TODO: use this feature more generally?
     out <- x$family$prior(dpar_class, link = link)
     if (!is.null(out)) {
@@ -1546,6 +1554,8 @@ validate_special_prior.btl <- function(x, prior, allow_autoscale = TRUE, ...) {
           "allowed when using special priors for the whole class."
         )
       }
+      # the brms function may not be in the search path of formula env
+      sub_prior <- paste0("brms::", sub_prior)
       tmp <- attributes(eval2(sub_prior, environment(x$formula)))
       tmp$autoscale <- isTRUE(tmp$autoscale) && allow_autoscale
       special[[sc]] <- tmp
@@ -1796,12 +1806,6 @@ prior_summary.brmsfit <- function(object, all = TRUE, ...) {
     prior <- prior[nzchar(prior$prior), ]
   }
   prior
-}
-
-#' @export
-default_prior.brmsfit <- function(object, ...) {
-  # just in case people try to apply default_prior to brmsfit objects
-  prior_summary.brmsfit(object, ...)
 }
 
 #' Checks if argument is a \code{brmsprior} object